No need for "base"
